refactor sellIn to int type

+++ src/test/java/cn/xpbootcamp/gildedrose/Goods.java
@@ -20,7 +20,7 @@ public class Goods {
         return quality;
     }
 
-    public double getSellIn() {
+    public int getSellIn() {
         return sellIn;
     }
 }
